video_scheduler/requirements.txt
fastapi
uvicorn
redis
pyyaml
requests
